rag-ingester/pyproject.toml
[project]
name = "rag-ingester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "chromadbx>=0.1.0",
    "fastapi[standard]>=0.116.1",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.6",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "requests>=2.32.4",
]
